test: move the known-defect tests to the specs that own the behaviour
known-defects-spec.ts organised tests by when the bugs were found rather
than by what the code does, so nobody editing Vector.compare or the coercion
layer would ever look there, and it duplicated homes that already existed.
Each test moves to the spec for the feature it covers:

  D1 multi-key ordering        data-frame-spec.ts ->orderBy
  D2 nil ordering (frame)      data-frame-spec.ts ->orderBy
  D2 the comparator itself     vector/vector-spec.ts ->compare
  D3 grapheme reversal         function-configs/string/reverse-spec.ts (new)
  D4 numeric coercion          type-coercion-spec.ts coerceToNumberType
                               (root cause already in core-utils)

reverse-spec.ts follows the split toGeoPoint and parseJSON use: the simple
cases stay in the function config's own `examples`, which function-configs-spec
runs for every function, and the unicode cases get a spec file because they
need escapes and a lone-surrogate check.

D5 (date coercion delegates to the JS Date parser) is not being fixed. Its
three cases are commented out at the foot of type-coercion-spec.ts with the
full write-up, rather than left failing: they assert the correct behaviour,
so a permanently red suite would just train people to ignore it.

No coverage is lost - the data-mate suite goes from 2443 passing and 3
failing to 2456 passing and none failing.
